add lua test script, minor fix to api files
