+++ modules/swagger-codegen/src/main/java/io/swagger/codegen/languages/LuaClientCodegen.java
@@ -181,11 +181,11 @@ public class LuaClientCodegen extends DefaultCodegen implements CodegenConfig {
 
     @Override
     public String apiFileFolder() {
-        return outputFolder + File.separator + "api" + File.separator;
+        return outputFolder + File.separator + packageName + File.separator + "api" + File.separator;
     }
 
     public String modelFileFolder() {
-        return outputFolder + File.separator + "model" + File.separator;
+        return outputFolder + File.separator + packageName + File.separator + "model" + File.separator;
     }
 
     @Override
