+++ modules/swagger-codegen/src/main/resources/lua/api.mustache
@@ -8,12 +8,13 @@ local dkjson = require "dkjson"
 local basexx = require "basexx"
 
 -- model import
-local {{{packageName}}}_{{classname}} = require "{{{packageName}}}.api.{{classname}}"
+-- TODO fix model import: local {{{packageName}}}_{{classname}} = require "{{{packageName}}}.model.{{classname}}"
 
-local {{{packageName}}}= {}
-local {{{packageName}}}_mt = {
+
+local {{{classname}}}= {}
+local {{{classname}}}_mt = {
 	__name = "{{{classname}}}";
-	__index = {{{packageName}}};
+	__index = {{{classname}}};
 }
 
 local function new_{{classname}}(host, basePath, schemes)
@@ -31,7 +32,7 @@ local function new_{{classname}}(host, basePath, schemes)
 		http_password = nil;
 		api_key = {};
 		access_token = nil;
-	}, {{{packageName}}}_mt)
+	}, {{{classname}}}_mt)
 end
 
 {{#operation}}
@@ -79,7 +80,7 @@ function {{classname}}:{{operationId}}({{#allParams}}{{paramName}}{{#hasMore}}, 
 	{{#isApiKey}}
 	{{#isKeyInHeader}}
 	-- api key in headers '{{keyParamName}}'
-	req.headers:upsert("{{{name}}}", api_key['{{{keyParamName}}}'])
+	req.headers:upsert("{{{name}}}", self.api_key['{{{keyParamName}}}'])
 	{{/isKeyInHeader}}
 	{{#isKeyInQuery}}
 	-- TODO: api key in query '{{keyParamName}}'
@@ -149,4 +150,8 @@ function {{classname}}:{{operationId}}({{#allParams}}{{paramName}}{{#hasMore}}, 
 end
 
 {{/operation}}
+return {
+    new = new_{{classname}};
+}
 {{/operations}}
+
