+++ samples/client/petstore/lua/.swagger-codegen/VERSION
@@ -1,1 +1,1 @@
-2.3.0-SNAPSHOT
+2.4.0-SNAPSHOT
