+++ samples/client/petstore/lua/test.lua
@@ -1,0 +1,12 @@
+local petstore_client = require "petstore.api.pet_api"
+local petstore_client_pet = require "petstore.model.pet"
+
+local my_pet_http_api = petstore_client.new("petstore.swagger.io", "/v2", {"http"})
+
+local my_pet = my_pet_http_api:get_pet_by_id(4)
+for k,v in pairs(my_pet) do
+	print(k,v)
+end
+
+local my_new_pet = petstore_client_pet.new("Mr. Barks")
+assert(my_pet_http_api:addPet(my_new_pet))
